move InvitationResource to v1
